This commit contains updated project with all models, dtos, repos and some implemented services

+++ checkit/src/main/java/com/example/checkit/dto/PreOrderDto.java
@@ -1,5 +1,6 @@
 package com.example.checkit.dto;
 
+
 import lombok.Data;
 import lombok.experimental.Accessors;
 
@@ -7,13 +8,13 @@ import lombok.experimental.Accessors;
 @Accessors(chain = true)
 public class PreOrderDto {
 
-    private Integer id;
+    private Long id;
 
     private String clientName;
 
     private String clientPhoneNumber;
 
-    private Long tripDistance;
+    private Float tripDistance;
 
     private Float deliveryCost;
